+++ CHANGELOG.md
@@ -4,14 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.10.2] - 2024-01-19
+- Add a feature in `syndicate update` command to update `logs_expiration`
+parameter in `cloudWatch` logs group. For setting `logs_expiration`, refer to
+`syndicate_aliases`.
+
 # [1.10.1] - 2024-01-17
 - Change from `error` to `warning` if integer not provided in allowed values for
 `logs_expiration` values and set default to `30`.
 
 # [1.10.0] - 2024-01-16
 - Add parameter `logs_expiration` to `syndicate_aliases.yml` and `lambda_config.json`. 
-The default value is set to "30 days". To ensure the logs never expire, set the
-value to: `logs_expiration: 0`.
+The default value is set to "30 days". To ensure the logs expire after 10 years,
+set the value to: `logs_expiration: 0`.
 
 # [1.9.9] - 2024-01-15
 - Fix dynamodb table capacity mode recognition during update